refactored codebase and implemented suggestions from sergioisidoro
