--- MANIFEST.in
@@ -1,3 +1,0 @@
-include *.txt
-include README.rst
-recursive-include django_webp *.py *.html *.png
