--- Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-ipython = "*"
-ipdb = "*"
-
-[packages]
-django = "*"
-pillow = "*"
-httpagentparser = "*"
-
-[requires]
-python_version = "3.8"
