+++ pyproject.toml
@@ -1,0 +1,51 @@
+[build-system]
+requires      = ["setuptools>=61.0.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "django-webp"
+version = "2.0.1"
+description = "Returns a webp image instead of jpg, gif or png to browsers"
+readme = "README.rst"
+requires-python = ">=3.5"
+keywords = ["django", "webp", "python"]
+authors = [
+    { name = "Andre Farzat", email = "andrefarzat@gmail.com" },
+    { name = "Daniel Opara", email = "daniel.opara@tufts.edu" }
+]
+maintainers = [
+  { name = "Andre Farzat", email = "andrefarzat@gmail.com" },
+  { name = "Daniel Opara", email = "daniel.opara@tufts.edu" }
+]
+license = { file = "LICENSE" }
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+  "Intended Audience :: Developers",
+
+  "Topic :: Software Development :: Libraries :: Python Modules",
+
+  "License :: OSI Approved :: MIT License",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+]
+
+dependencies = [
+  "asgiref>=3.5.0",
+  "django>=4.0.3",
+  "pillow>=9.0.1",
+  "sqlparse>=0.4.2",
+]
+
+[project.optional-dependencies]
+dev = ["pip-tools", "pytest", "black"]
+
+[project.urls]
+Homepage = "http://pypi.python.org/pypi/django-webp/"
+
+[tool.setuptools]
+include-package-data = true
+
+[tool.setuptools.packages.find]
+namespaces = true
+where = ["src"]
+
