+++ setup.py
@@ -7,13 +7,13 @@ with open('README.rst') as file:
 
 setuptools.setup(
     name='django-webp',
-    version='2.0.0',
+    version='2.0.1',
     author=u'Andre Farzat',
     author_email='andrefarzat@gmail.com',
     packages=setuptools.find_packages(),
     url='http://pypi.python.org/pypi/django-webp/',
     license='MIT',
-    description='Returns a webp image instead of jpg, gif or png to browsers which have support',
+    description='Returns a webp image instead of jpg, gif or png to browsers',
     long_description=long_description,
     install_requires=open('requirements.txt').readlines(),
     include_package_data=True,
