+++ src/django_webp/management/commands/clean_webp_images.py
