+++ src/django_webp/templatetags/webp.py
