+++ tests/mediafiles/inside/folder/python3.png
