+++ tests/test_WEBPImageConverter.py
