+++ tests/testapp/django_webp/python.png
