+++ tests/testapp/static/django_webp/python.png
