+++ tests/testapp/templates/django_webp/index.html
